Snake game in C++
A simple snake game in c++ in which the idea is not to erase the complete screen and print newly. On snake's movement it just erase the ending character (the tail of the snake) due to which the screen does not seem blinking.
I have also added a classic mode where the snake gets killed when hit to the wall.
It automatically adjusts if the window size changes.
It only works in Windows OS because of the usage of graphics.h and windows specific functions.